Componentのサイズが肥大化したので分割した
