+++ .prettierrc
@@ -1,0 +1,3 @@
+{
+    "printWidth": 120
+}
