+++ src/components/calender.tsx
@@ -1,8 +1,21 @@
 import React, { useState } from "react";
-import { format, getDate, getDay, eachDayOfInterval, endOfWeek, eachWeekOfInterval,addMonths,subMonths,startOfMonth,endOfMonth,isSameMonth,isSameDay } from "date-fns";
+import {
+  format,
+  getDate,
+  getDay,
+  eachDayOfInterval,
+  endOfWeek,
+  eachWeekOfInterval,
+  addMonths,
+  subMonths,
+  startOfMonth,
+  endOfMonth,
+  isSameMonth,
+  isSameDay
+} from "date-fns";
 import styled from "styled-components";
-import { blue, red ,pink } from "@material-ui/core/colors";
-import { Paper,Button, Grid, Typography, Table, TableBody, TableCell, TableHead, TableRow } from "@material-ui/core";
+import { Paper, Button, Grid, Typography, Table, TableBody, TableCell, TableHead, TableRow } from "@material-ui/core";
+import CalendarTableCell from "./calendarTableCell";
 
 // Material-UI のButtonをカッコで囲んで、styled の引数にしてCSS in JSを実現している.
 const StyledButton = styled(Button)`
@@ -29,66 +42,13 @@ const StyleTableHeadCell = styled(TableCell)`
   text-align: center;
 `;
 
-interface Props {
-  wday: Number;
-  isTargetMonth: Boolean;
-  istargetday: boolean;
-  children: any;
-}
-
-// Propsで扱っている情報を引数として日付として取得し、表示用のstyled-componentsに格納
-function CalendarTableCell(props: Props) {
-  const { wday, isTargetMonth, istargetday, children, ...other } = props;
-  return (
-    <StyleTableCell
-      wday={props.wday}
-      isTargetMonth={props.isTargetMonth}
-      istargetday={props.istargetday}
-  >{props.children}</StyleTableCell>
-  );
-}
-
-// 引数に応じてCellの色を変化させる。対象月、本日、土日と平日などで色分け
-const StyleTableCell = styled(TableCell)<{
-  wday: Number;
-  isTargetMonth: Boolean;
-  istargetday: boolean;
-}>`
-  text-align: center;
-  color: ${({ wday, isTargetMonth }): string => {
-    if (isTargetMonth) {
-      switch (wday) {
-        case 0: // Sunday
-          return red[500];
-        case 6: // Saturday
-          return blue[500];
-        default:
-          return "black";
-      }
-    } else {
-      // previous or next month
-      switch (wday) {
-        case 0: // Sunday
-          return red[200];
-        case 6: // Saturday
-          return blue[200];
-        default:
-          return "black";
-      }
-    }
-  }};
-  background: ${(istargetday): string => (istargetday ? pink[50] : blue[500])};
-`;
-
 // カレンダーに表示するための一月分のArrayを返す関数
 const getCalendarArray = (date: Date): Date[][] => {
   const sundays = eachWeekOfInterval({
     start: startOfMonth(date),
     end: endOfMonth(date)
   });
-  return sundays.map(sunday =>
-    eachDayOfInterval({ start: sunday, end: endOfWeek(sunday) })
-  );
+  return sundays.map(sunday => eachDayOfInterval({ start: sunday, end: endOfWeek(sunday) }));
 };
 
 // カレンダーを表示するReactのFunctionComponent
@@ -103,30 +63,17 @@ const calendar: React.FC = () => {
       <StyledPaper>
         <Grid container justify="space-between">
           <Grid item>
-            <StyledButton
-              type="button"
-              onClick={(): void =>
-                setTargetDate(current => subMonths(current, 1))
-              }
-            >
+            <StyledButton type="button" onClick={(): void => setTargetDate(current => subMonths(current, 1))}>
               前の月
             </StyledButton>
           </Grid>
           <Grid item>
-            <StyledButton
-              type="button"
-              onClick={(): void => setTargetDate(new Date())}
-            >
+            <StyledButton type="button" onClick={(): void => setTargetDate(new Date())}>
               今月
             </StyledButton>
           </Grid>
           <Grid item>
-            <StyledButton
-              type="button"
-              onClick={(): void =>
-                setTargetDate(current => addMonths(current, 1))
-              }
-            >
+            <StyledButton type="button" onClick={(): void => setTargetDate(current => addMonths(current, 1))}>
               次の月
             </StyledButton>
           </Grid>
